CamundaTaskWaInitiationTest now asserting and passing, resultsMatchUsingNameKey() moved to new class CamundaUtils
