+++ src/test/java/uk/gov/hmcts/probate/dmn/CamundaTaskWaConfigurationTest.java
@@ -23,7 +23,6 @@ import java.util.stream.Stream;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static uk.gov.hmcts.probate.DmnDecisionTable.WA_TASK_CONFIGURATION_PROBATE;
 import static uk.gov.hmcts.probate.dmnutils.CamundaTaskConstants.CASE_MANAGEMENT_CATEGORY;
 import static uk.gov.hmcts.probate.dmnutils.CamundaTaskConstants.CASE_NAME;
@@ -35,6 +34,7 @@ import static uk.gov.hmcts.probate.dmnutils.CamundaTaskConstants.LOCATION;
 import static uk.gov.hmcts.probate.dmnutils.CamundaTaskConstants.REGION;
 import static uk.gov.hmcts.probate.dmnutils.CamundaTaskConstants.ROLE_CATEGORY;
 import static uk.gov.hmcts.probate.dmnutils.CamundaTaskConstants.WORK_TYPE;
+import static uk.gov.hmcts.probate.dmnutils.CamundaUtils.resultsMatchUsingNameKey;
 
 class CamundaTaskWaConfigurationTest extends DmnDecisionTableBaseUnitTest {
 
@@ -126,78 +126,4 @@ class CamundaTaskWaConfigurationTest extends DmnDecisionTableBaseUnitTest {
                 && (now.isEqual(result) || now.isAfter(result));
     }
 
-    private void resultsMatchUsingNameKey(List<Map<String, Object>> results, List<Map<String, Object>> expectation) {
-        assertThat(results.size(), is(expectation.size()));
-
-        for (Map<String, Object> expectedEntry : expectation) {
-            String expectedName = (String) expectedEntry.get("name");
-            Map<String, Object> resultEntry = results.stream()
-                    .filter(result -> expectedName.equals(result.get("name")))
-                    .findFirst()
-                    .orElseThrow(() -> new AssertionError("No result found for name: " + expectedName));
-
-            for (String key : expectedEntry.keySet()) {
-                assertEquals(expectedEntry.get(key), resultEntry.get(key),
-                        "Mismatch for key: " + key + " in entry with name: " + expectedName);
-            }
-        }
-    }
-
-    @Test
-    void shouldMatchResultsWhenAllEntriesMatchByName() {
-        List<Map<String, Object>> results = List.of(
-                Map.of("name", "task1", "value", "value1", "canReconfigure", true),
-                Map.of("name", "task2", "value", "value2", "canReconfigure", false)
-        );
-
-        List<Map<String, Object>> expectation = List.of(
-                Map.of("name", "task1", "value", "value1", "canReconfigure", true),
-                Map.of("name", "task2", "value", "value2", "canReconfigure", false)
-        );
-
-        resultsMatchUsingNameKey(results, expectation);
-    }
-
-    @Test
-    void shouldThrowErrorWhenResultIsMissingForExpectedName() {
-        List<Map<String, Object>> results = List.of(
-                Map.of("name", "task1", "value", "value1", "canReconfigure", true)
-        );
-
-        List<Map<String, Object>> expectation = List.of(
-                Map.of("name", "task1", "value", "value1", "canReconfigure", true),
-                Map.of("name", "task2", "value", "value2", "canReconfigure", false)
-        );
-
-        assertThrows(AssertionError.class, () -> resultsMatchUsingNameKey(results, expectation));
-    }
-
-    @Test
-    void shouldThrowErrorWhenValuesDoNotMatchForSameName() {
-        List<Map<String, Object>> results = List.of(
-                Map.of("name", "task1", "value", "value1", "canReconfigure", true),
-                Map.of("name", "task2", "value", "wrongValue", "canReconfigure", false)
-        );
-
-        List<Map<String, Object>> expectation = List.of(
-                Map.of("name", "task1", "value", "value1", "canReconfigure", true),
-                Map.of("name", "task2", "value", "value2", "canReconfigure", false)
-        );
-
-        assertThrows(AssertionError.class, () -> resultsMatchUsingNameKey(results, expectation));
-    }
-
-    @Test
-    void shouldThrowErrorWhenResultsSizeDoesNotMatchExpectationSize() {
-        List<Map<String, Object>> results = List.of(
-                Map.of("name", "task1", "value", "value1", "canReconfigure", true)
-        );
-
-        List<Map<String, Object>> expectation = List.of(
-                Map.of("name", "task1", "value", "value1", "canReconfigure", true),
-                Map.of("name", "task2", "value", "value2", "canReconfigure", false)
-        );
-
-        assertThrows(AssertionError.class, () -> resultsMatchUsingNameKey(results, expectation));
-    }
 }
